LIDE-004: Finalizing challenge 101 for merge
Moved example solution into a solution directory
Removed solution from source file; Left placeholders ("put your code
here")

Changes to be committed:
	modified:   101_source.c
	new file:   solution/101_example.c
